T7.35 int -> long

+++ src/main/java/com/gtnewhorizons/galaxia/registry/celestial/CelestialAsset.java
@@ -223,7 +223,7 @@ public abstract class CelestialAsset implements Buildable, IDistributedInventory
         dirty = false;
     }
 
-    public abstract long updateContents(InventoryKey item, int delta, boolean sync);
+    public abstract long updateContents(InventoryKey item, long delta, boolean sync);
 
     /// ----------------------------------------------------------------------------------
     /// Inventory Bounds
